pay the circular dependency tax
