+++ editor/src/components/editor/store/editor-state.ts
@@ -191,7 +191,6 @@ import { emptyImports } from '../../../core/workers/common/project-file-utils'
 import type { CommentFilterMode } from '../../inspector/sections/comment-section'
 import type { Collaborator } from '../../../core/shared/multiplayer'
 import type { OnlineState } from '../online-status'
-import type { RenderedAt } from '../../inspector/sections/component-section/data-reference-cartouche'
 
 const ObjectPathImmutable: any = OPI
 
@@ -2172,6 +2171,29 @@ export function syntheticNavigatorEntry(
     elementPath: elementPath,
     childOrAttribute: childOrAttribute,
   }
+}
+
+interface RenderedAtPropertyPath {
+  type: 'element-property-path'
+  elementPropertyPath: ElementPropertyPath
+}
+
+interface RenderedAtChildNode {
+  type: 'child-node'
+  parentPath: ElementPath
+  nodeUid: string
+}
+
+export type RenderedAt = RenderedAtPropertyPath | RenderedAtChildNode
+
+export function renderedAtPropertyPath(
+  elementPropertyPath: ElementPropertyPath,
+): RenderedAtPropertyPath {
+  return { type: 'element-property-path', elementPropertyPath: elementPropertyPath }
+}
+
+export function renderedAtChildNode(parentPath: ElementPath, nodeUid: string): RenderedAtChildNode {
+  return { type: 'child-node', parentPath: parentPath, nodeUid: nodeUid }
 }
 
 export interface DataReferenceNavigatorEntry {
