+++ editor/src/components/editor/store/store-deep-equality-instances.ts
@@ -358,6 +358,7 @@ import type {
   DataReferenceNavigatorEntry,
   GithubUser,
   PullRequest,
+  RenderedAt,
 } from './editor-state'
 import {
   trueUpGroupElementChanged,
@@ -368,6 +369,8 @@ import {
   renderPropValueNavigatorEntry,
   dataReferenceNavigatorEntry,
   newGithubData,
+  renderedAtPropertyPath,
+  renderedAtChildNode,
 } from './editor-state'
 import {
   editorStateNodeModules,
@@ -605,11 +608,6 @@ import type {
 } from '../../custom-code/internal-property-controls'
 import type { OnlineState } from '../online-status'
 import { onlineState } from '../online-status'
-import type { RenderedAt } from '../../inspector/sections/component-section/data-reference-cartouche'
-import {
-  renderedAtChildNode,
-  renderedAtPropertyPath,
-} from '../../inspector/sections/component-section/data-reference-cartouche'
 
 export function ElementPropertyPathKeepDeepEquality(): KeepDeepEqualityCall<ElementPropertyPath> {
   return combine2EqualityCalls(
@@ -705,12 +703,43 @@ export const SyntheticNavigatorEntryKeepDeepEquality: KeepDeepEqualityCall<Synth
     syntheticNavigatorEntry,
   )
 
+export const RenderedAtKeepDeepEquality: KeepDeepEqualityCall<RenderedAt> = (
+  oldValue,
+  newValue,
+) => {
+  switch (oldValue.type) {
+    case 'element-property-path':
+      if (oldValue.type === newValue.type) {
+        return combine1EqualityCall(
+          (r) => r.elementPropertyPath,
+          ElementPropertyPathKeepDeepEquality(),
+          renderedAtPropertyPath,
+        )(oldValue, newValue)
+      }
+      break
+    case 'child-node':
+      if (oldValue.type === newValue.type) {
+        return combine2EqualityCalls(
+          (r) => r.parentPath,
+          ElementPathKeepDeepEquality,
+          (r) => r.nodeUid,
+          StringKeepDeepEquality,
+          renderedAtChildNode,
+        )(oldValue, newValue)
+      }
+      break
+    default:
+      assertNever(oldValue)
+  }
+  return keepDeepEqualityResult(newValue, false)
+}
+
 export const DataReferenceNavigatorEntryKeepDeepEquality: KeepDeepEqualityCall<DataReferenceNavigatorEntry> =
   combine4EqualityCalls(
     (entry) => entry.elementPath,
     ElementPathKeepDeepEquality,
     (entry) => entry.renderedAt,
-    createCallWithTripleEquals(), // TODO
+    RenderedAtKeepDeepEquality,
     (entry) => entry.surroundingScope,
     ElementPathKeepDeepEquality,
     (entry) => entry.childOrAttribute,
@@ -755,37 +784,6 @@ export const SlotNavigatorEntryKeepDeepEquality: KeepDeepEqualityCall<SlotNaviga
     StringKeepDeepEquality,
     (elementPath, prop) => ({ type: 'SLOT', elementPath: elementPath, prop: prop }),
   )
-
-export const RenderedAtKeepDeepEquality: KeepDeepEqualityCall<RenderedAt> = (
-  oldValue,
-  newValue,
-) => {
-  switch (oldValue.type) {
-    case 'element-property-path':
-      if (oldValue.type === newValue.type) {
-        return combine1EqualityCall(
-          (r) => r.elementPropertyPath,
-          ElementPropertyPathKeepDeepEquality(),
-          renderedAtPropertyPath,
-        )(oldValue, newValue)
-      }
-      break
-    case 'child-node':
-      if (oldValue.type === newValue.type) {
-        return combine2EqualityCalls(
-          (r) => r.parentPath,
-          ElementPathKeepDeepEquality,
-          (r) => r.nodeUid,
-          StringKeepDeepEquality,
-          renderedAtChildNode,
-        )(oldValue, newValue)
-      }
-      break
-    default:
-      assertNever(oldValue)
-  }
-  return keepDeepEqualityResult(newValue, false)
-}
 
 export const NavigatorEntryKeepDeepEquality: KeepDeepEqualityCall<NavigatorEntry> = (
   oldValue,
